refactor(cloudflare/vectorize): extract shared HTTP client builder
Move the SearchIndexClient construction + shape adapters into unexported
SearchIndexHttpClient.ts, parameterized by an injectable SearchIndexAuth
({ authorize, accountId }) — same convention as KV/R2/Tunnel/AI. A future
token-scoped SearchIndexHttp layer reuses the builder; SearchIndexLocal is
now just the layer + auth.
